Splitting test suites into different files
